+++ src/htmx/index.js
@@ -1,8 +1,9 @@
 const HtmlRenderer = require('./HtmlRenderer');
 const LiveHtmlServer = require('./LiveHtmlServer');
 const LivePage = require('./LivePage');
-const { html } = require('./Html');
+const { attribute, codeBlock, each, html, safeUrl: url } = require('./Html');
 const { action, page, state, view } = require('./metadata');
 const { start } = require('./start');
+const { exportStatic } = require('./StaticExporter');
 
-module.exports = { action, html, HtmlRenderer, LiveHtmlServer, LivePage, page, start, state, view };
+module.exports = { action, attribute, codeBlock, each, exportStatic, html, HtmlRenderer, LiveHtmlServer, LivePage, page, start, state, url, view };